[BaseRepositoryBenchmarkTest] - replay BaseRepositoryBenchmarkTest with 20 iterations - only average time benchmark
(results in 'jmh/BaseRepositoryBenchmarkTest.results')

- with single bean :
  > find_all_with_single_bean_BaseRepository_findAll
  > find_all_with_single_bean_SimpleJpaRepository_findAll

The results are similar in average time.

- with iterations of list :
  > find_all_with_list_BaseRepository_findAll
  > find_all_with_list_SimpleJpaRepository_findAll

The results are significativelly different, but totally normal because of the 'BaseRepository.findAll(Iterable iterable)' method's purpose.

Note : if spring had implemented it, we would have one query to find all the bean list instead to have one query/bean in a loop as done in :
BaseRepositoryBenchmarkTest.find_all_with_list_SimpleJpaRepository_findAll
> The results would be very probably quite different (i don't worry about :)

This test is only here to approuve the interest of a custom method findAll taking an entity set, instead of one bean only.