[ExtendedJpaRepository] - fork of SimpleJpaRepository reviewed to extend specifications
TOPIC : How to extend SimpleJpaRepository specifications from a custom repository implementation ? SimpleJpaRepository is hardly extendable, because of restrictions to read-only operations in criteria uses done in this class.

> Specification.class / SpecificationComposition.class and all methods implemented in SimpleJpaRepository use CriteriaQuery type to generate criteria 'where' clauses.
The fact to not use a common parent interface ancestor of CriteriaQuery.class like CommonAbstractCriteria.class (JPA api) doesn't allow to add generic behaviours handling for write operations using criteria (to keep using criteria as default query builder)

In fact if we want to add a method deleteAll(Iterable) respecting the specification pattern used in SimpleJpaRepository, we cannot because the only actual criteria type used is CriteriaQuery type so SimpleJpaRepository  doesn't handle write operations with criteria.

Certain method in SimpleJpaRepository do write operations but they only use the entityManager's methods (so the scope is allways one unit entity) :

example : save, saveAll, delete, deleteById(), deleteAll()

Pehaps these methods can be optimized, to avoid looping with multiple queries.
(For example deleteAll is a loop with one EntityManager.remove() applied for each entity !)

> going to test a re-write of these methods to avoid multiple queries, and compare the perfs in future.

> SimpleJpaRepository doesn't use CriteriaDelete class to be able to delete multiple records with one criteria query, or CriteriaUpdate class to update multiple records with one query.

> ExampleSpecification / ByIdsSpecification are inner classes of SimpleJpaRepository

The actual SimpleJpaRepository doesn't seem to be easily extendable, we need to do a lot of things and finally avoid default Specification interface methods (see default methods in ExtendedJpaRepository.class), if we want to add write operations in new specifications, because of prototyping chosen in Specification.class.

If we want to add a new specification to handle writes using criteria, we cannot just create a custom specification class inheriting of Specification interface and use it with SimpleJpaRepository or a base jpa implementation, except for read operations therefore using CriteriaQuery..

Then, because of certain protected resources used by SimpleJpaRepository, we are not allowed to add one spec with the flexibility of using the implementations of parent classes.

An important thing is that the CriteriaQuery type taken as parameter of method 'Specification.toPredicate' and used in SpecificationComposition.class is never used in the existing implementations.
(The code looks to have been generated at some points)

Note : to open the implementation, perhaps specification's classes could be externalized of the SimpleJpaRepository class for coherence with the visibility of Specification interface and for flexibility. These specifications classes are private static and inner classes of SimpleJpaRepository at this moment.

Not a troll, not a critic because i love the magic of spring-data for sure and SimpleJpaRepository is probably going to change in the future.

AND : On another hand, we can simply working without the specification principle and directly write a classical method implementing our needs using Criteria or Hql as we used to do in a classical custom repository implementation.

But quizz of a smart way to keep using SimpleJpaRepository and add a custom specification with write operations using super classes implemented ?